refactor: rename events.js to notificationEvents.js
